policies and middleware in controllers

+++ app/Http/Controllers/App/MediaController.php
@@ -5,13 +5,23 @@ namespace App\Http\Controllers\App;
 use App\Http\Controllers\Controller;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
 
 class MediaController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index(Project $project)
     {
+        if (!Auth::user()->canAny(['projects.create.' . $project->id, 'projects.update.' . $project->id])){
+            abort(403);
+        }
+
         return response()->json([
             'media_images' => $project->getMedia('images')->map(function (Media $media) {
                 return [
@@ -36,6 +46,10 @@ class MediaController extends Controller
 
     public function store(Request $request, Project $project)
     {
+        if (!Auth::user()->canAny(['projects.create.' . $project->id, 'projects.update.' . $project->id])){
+            abort(403);
+        }
+
         $this->validate($request, [
             'media' => ['required', 'mimes:jpg,png,mp4,wav,gif', 'max:10240']
         ]);
@@ -54,6 +68,10 @@ class MediaController extends Controller
 
     public function destroy(Request $request, Project $project)
     {
+        if (!Auth::user()->canAny(['projects.create.' . $project->id, 'projects.update.' . $project->id])){
+            abort(403);
+        }
+
         $project->deleteMedia($request->media_id);
     }
 }
